Added files via upload
Given n filesystems a common subtree is created consisting of files and directories which are present in majority filesystems. A filesystem is present in the common subtree if it is present in majority of filesystems(ceil(n/2)) with the same relative pathname and content in the filesystems. A directory is common if it is present in majority of filesystems and its subdirectories and files are common in the same filesystems in which it is common.